resource-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose==3.3.0
pydantic==2.9.2
httpx==0.27.2
kafka-python==2.0.2
